Beginning of the new, JavaScript-ed website.
Layout for the home page is done; now just need to enable the links and add
the content. The plan is to slide the menu up (removing the picture window)
just below the title, and add the content underneath for each page. Each will
be one page(ish) of content, with links to other pages for the big topics
(SCRABBLE StuCo, Resume, big projects, etc.).