Links and section navigation added.
The links to sections are now enabled and looking good (still no subsections
for Projects and More, as the content hasn't been added yet). Dummy content
in each section for now. Still need to address the backed-up queueing issue
with the slide animations (for the picture window, menu option-lists, and
section navigation(? almost passable w/o)). Then, section content needs to
be filled in, and links to subsections added (this may screw up animations).