Massive changes. New version of site complete
Directory structure was changed a lot; lots of files renamed. All the content
was added, including pictures, additional styling, day/night mode, everything
all finished. Ready to upload. The updatesite.sh script now just automatically
scps everything in each of the 'live' directories (in ., img, and scrabble
right now) to the appropriate folders on Andrew; much more convenient. Just
remember to keep everything you need for the live site (and only that) in the
live folders, and not to create any subdirectories within them (directories
will be ignored by scp, so it's not a huge problem if they exist). Launching!